sentiment-app/requirements.txt
# --- FastAPI Backend ---
fastapi==0.110.1
uvicorn[standard]==0.29.0
boto3==1.34.84
python-dotenv==1.0.1

# --- Streamlit Frontend ---
streamlit==1.32.2
requests==2.31.0

# --- Optional for JSON preview, pandas if needed ---
pandas==2.2.2
